requirements.txt
easyocr>=1.7.1
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
potrace>=0.1.0
segment-anything>=1.0.0
torch>=2.0.0
torchvision>=0.15.0 